Move cross/longitudinal to SubjectDerivedNamespace:
