+++ nacc_attribute_deriver/attributes/nacc/modules/uds/apoe_derived.py
@@ -1,11 +1,11 @@
 from nacc_attribute_deriver.attributes.attribute_collection import AttributeCollection
-from nacc_attribute_deriver.attributes.base.uds_namespace import UDSNamespace
+from nacc_attribute_deriver.attributes.base.namespace import SubjectDerivedNamespace
 from nacc_attribute_deriver.symbol_table import SymbolTable
 
 
 class APOEDerivedAttributeCollection(AttributeCollection):
     def __init__(self, table: SymbolTable):
-        self.__uds = UDSNamespace(table)
+        self.__subject_derived = SubjectDerivedNamespace(table)
 
     def _create_naccapoe(self) -> int:
-        return self.__uds.get_cross_sectional_value("naccapoe")
+        return self.__subject_derived.get_cross_sectional_value("naccapoe")
